[RPP Insight] Use event ref in Network dependency tree
Use the existing EventRef component so it can be consistent with other insights. Also can have the 'click' to revel in flame chart feature.

Synced with Yuin, and we decide to remove the underline for the tree view, so added an override-able css variable.

Bug: 372897712
Change-Id: I48f14acc356f0188dae123004dd04e1ccf566c1f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6305801
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>