Add squiggly line support for <select> issues in DOM tree
This CL adds a separate utility collection for issues in the Elements
panel, which will be used for both the DOM tree and the a11y tree. This utility collection is responsible for adapting various types of issues and return "normalized" result for the Elements panel to display. This way, we keep the DOM tree agnostic to the adaption and the types of issues, hence the relaxation of the issue type from GenericIssue to Issue.

For now, the state management of Elements panel-related issues stays in the Elements outline.

Bug: 378738916, 393245109
Change-Id: Ifbabc14b735cc0f7568e7a46bd7ea1807b52dc25
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6297955
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>