Close tooltip automatically if anchor is removed
The css solution works when scrolling/moving the anchor outside of the viewport.

The mutation observer is needed when the tooltip can't be nested as a child of the anchor and is removed from the DOM while the tooltip is still open. A use-case for this is the editing of a css selector, while the specificity tooltip is still visible, see http://crrev/c/6298060.

Demo: https://screencast.googleplex.com/cast/NDgzNjA0OTc1MzczNTE2OHxlYzllMmU0MC0yMA

Change-Id: I3092910abb8eb397cd2ba0fe0d01dd721ecfd4ff
Bug: 399801002
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6308773
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Marcel Pütz <marcelpuetz@chromium.org>