RPP: add `getMainThreadActivity` function to Insights AI
This CL enables the LLM to query for relevant main thread activity for
the given insight.

Note that we may need to experiment with alternative versions of this
that are leaner in order to avoid eating up context windows, but this is
a good starting point.

Bug: 394552594
Change-Id: I2017f190f783743f813abea57e9dfca92c9f6855
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6275957
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>