[build] Remove 'devtools_use_remoteexec' GN arg
This was an initial experiment to make ts_library work with goma. We
got as far as building leaf modules remotely but getting the rest to
work was deemed infeasible.

This CL removes the prototype.

Note that this is different from siso + use_remoteexec. That works
silently under the hood and only relies on inputs/outputs being
defined correctly. It doesn't require special support in our
ts_library.py.

R=pfaffe@chromium.org

Bug: 40153338
Change-Id: I6de5e30eeb74b80a2ec591dfff0f700a584a43a7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6309563
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>