AI: provide network summary to Perf Insights agent
Someone found a bug where their site broke the agent because it had 180
requests that the agent asked for details of and this quickly ate up the
available context window. This CL changes the architecture so the LLM is
expected to first get a summary of the network requests before then
being allowed to enquire about specific requests.

There is definitely more room to experiment here; we also could provide
the LLM with a list of "critical" requests based on heuristics we have,
rather than providing a list of all requests. This needs more
experimentation.

Fixed: 400909537
Change-Id: I0d224834010dad050411a0024d226ca1e22f3664
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6324575
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>