[AI Assistance] Collect source information
This CL adds a method to collect information about the source file
associated with the style rule we are taking a selector from. The
filename is not used in this CL so far but in a follow-up I will
integrate it with the ChangeManager.

Bug: 400875470
Change-Id: I58cd1799573497a5f36bab0d0d86b5f4971d1ae4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6328142
Reviewed-by: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>