[AI] Refactor AI settings
Since the settings structure is identical, there was code duplicated for the 2 existing AI settings and with the addition of the AI Generated Labels, it would have been a triplicate. There might be even more settings in the future.
Therefore, this cl refactors the settings to have reusable methods to render AiSettingParams objects.

The follow up cl will add the new setting for AI labels.

Bug: 399814051
Change-Id: I61112e38870040629f5584117d850a1314d8fd9c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6336215
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>