RPP: fix main thread lookup for Insights
To get the data for the `getMainThreadActivity` function that we pass to
the LLM, we were finding the main thread and using that. But it is
possible for a trace to have two main threads, and if we pick the wrong
one, the data for the LLM is incorrect.

This CL updates the code to try to find the right thread based on the
navigation associated with the insight, if it has it. This is a bit of a
stop-gap but should improve the situation. I have logged the bug below
to track the work to fix this properly in the trace engine; right now
the existence of two "main threads" is breaking some assumptions in the
data layer that means this work around is required.

Bug: 402658800
Change-Id: Ia8931facf51afb4fd483919b9a074f0c0cc0a762
Fixed: 402660533
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6346608
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>