RPP: sort process windows before setting them
In the screenshot of the bug below, notice how for one process for a
frame we end up setting its time range to invalid, negative numbers.
This is because we were not sorting the windows first, which is
required. We had an assumption that the timestamps of the trace events
come in time order, but that is clearly not always true, so this CL
updates the code to sort first to ensure we have them ordered correctly
before we then set the bounds for each process.

Bug: 402658800
Change-Id: I1b0a394f98cf4538f81cfa4debada97987655593
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6346911
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Nancy Li <nancyly@chromium.org>