[cleanup] Introduce `UI.UIUtils.openInNewTab()` helper.
This merges the logic for adding missing UTM parameters to links to
Google owned documentation sets into a single helper which we use
to open links that potentially contain documentation set references
in a new Chrome tab.

This way we reduce the chance of missing the `utm_source` or the
`utm_campaign` parameters in the future.

Fixed: 402611137
Change-Id: I1efc7ae90beb59dee12741470f8deae16e1c7fca
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6346913
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>