Use two worker threads during heap snapshot initialization
This change creates a secondary short-lived Worker to help out during
HeapSnapshot.initialize, and moves three of the existing initialization
steps to that secondary Worker. Those three steps are carefully chosen
to minimize the amount of data that must be copied and maximize the
ability to make progress on both threads at once. Roughly speaking, the
second thread takes on any steps that involve upward traversal through
the graph (using `retainingNodes` and friends). This approach requires
some more memory during initialization, because the following data must
be duplicated for the second thread:

* For each edge, the node ordinal it points to (4 bytes per edge)
* For each node, the index of its first outgoing edge (4 bytes per node)
* For each node, its shallow size (4 bytes per node)

However, the speedup is substantial. A 500 MB .heapsnapshot file takes
about 12 seconds to load with this change, versus 18 seconds without.

I've tried to structure the code in a way that prevents any possibility
of inconsistent behavior or data races. All steps on the secondary
thread clearly define their inputs and outputs, and the primary thread
doesn't observe any results from the secondary thread until it waits for
those results in installResultsFromSecondThread.

Bug: 42203857
Change-Id: Idc3604d127012ecb2e5daf910b7844727ca6c0af
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6339201
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>