Have InspectorView remove the Infobar instead of CookieControlsView
If CookieControlsView was responsible for listening for a
PrimaryPageChanged event and then removing the Infobar, the infobar
would not be removed if CookieControlsView was never created to
listen in the first place. This is a problem because the Infobar
might be created by the InspectorMain entrypoint which does not
guarantee that a CookieControlsView will be created as well.

By moving the listener to the InspectorView, we guarantee that the
Infobar is removed when the user reloads the page without needing
to open the Privacy and security panel.

Bug: 400962121
Change-Id: I612c4899e218e094fd31a3eda154ec69c20e8035
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6346989
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>