[RPP] Create row limit and aggregate helper
One strategy I considered was to add a `maxRows` property to the table
component. The overlays for the "N others" row would be aggregated from
the table rows beyond `maxRows`. However there are a couple drawbacks to
this strategy:
- We would be creating lit templates for table rows that don't actually get rendered.
- It isn't possible to fill the second column of the "N others" row with relevant data (e.g. total network payload of the rows not shown)

The strategy taken in this CL resolves the above concerns but is less
DRY.

Bug: None
Change-Id: Id6e4dde29712f92c867d605ec6239b772feabcaa
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6355097
Commit-Queue: Adam Raine <asraine@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>