AI: manually strip 5 backticks from LLM response in Insights agent
See bug for context, we sometimes see the LLM wrap its response in 5
backticks for no apparent reason. We're going to sync with the Aida
folks to see if we can understand why or fix this on their side, but for
now a fix is also to strip the 5 backticks if they are found in the
response.

Bug: 405054694
Change-Id: Ieb7499af142f7f7cba9ffea999729f49e1fbe536
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6375525
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Ergün Erdoğmuş <ergunsh@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>