[css value tracing] evaluate percentages in longhands
This adds support for evaluating <percentage> units in css value
tracing, but only for longhands. For shorthands, we need additional reasoning about which longhand the unit pertains to in order to understand whether it's relative to a width or a height.

Bug: 401213719
Change-Id: I823ef9d52bb12ab40e7bcabce770733ad36a51b7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6376104
Reviewed-by: Eric Leese <leese@chromium.org>
Auto-Submit: Philip Pfaffe <pfaffe@chromium.org>
Commit-Queue: Eric Leese <leese@chromium.org>