[Perf panel] Remove support for experimental server timings track
We did not receive clear signal about the usefulness of having a mechanism to display server timings in the client timeline.

Developers will still be able to see server timings in the details of network requests (as of https://crrev.com/c/6387100) and if needed, a server timing track could be injected using the extensibility API. (granted, a mechanisim to map server timestamps to the client clock needs to be implemented, but this doesn't seem to be a big constraint as devs are alreading doing so, like React).

Bug: 403223638
Change-Id: I10d80fe10598dcf7c3014cfc1fcbce03b6024fb5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6387102
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>