Define window.scheduler as non-optional
We used the wicg-task-scheduling types for `window.scheduler` but it
types it as optional - this is true for the general case but because we
are in Chrome we know it exists.

Unfortunately you cannot override the member from optional or not on
`Window`, so this CL removes the use of the
`@types/wicg-task-scheduling` package and reuses the code inline with
the extension to the `Window` interface making it always present.

Bug: none
Change-Id: I8399d3477f89f88e758ed31399c6256fcbd31de9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6393155
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Connor Clark <cjamcl@chromium.org>
Commit-Queue: Connor Clark <cjamcl@chromium.org>