RPP: support call tree generation (and AI labels) for hidden events
If the user turns on the "Show all events" experiment, they will see
events that are not listed in our default `visibleTypes` list. This was
problematic for the call tree generation which relies on that list to
generate the data to send to the AI for label generation and "Ask AI".

This CL now updates the implementation to filter for visible types only
when that experiment is disabled.

As a drive-by, I also wrapped the AICallTree generation in a rAF to
ensure the UI updates are snappy.

Fixed: 406004077
Change-Id: Iea80cc59771738389de77375828caee16e654f35
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6396679
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>