Reduce the number of casts to Node in preferTemplateLiterals
The problem is that ESLint defines a `parent` property on the estree Node. However specific estree Node subtypes, such as MemberExpression, have properties with estree Node subtypes, such as Identifier. The value in practice does have a `parent` property but this is not reflected in types.

This CL makes subrules unaware of the difference by exposing only the estree type to them and makes "core" libraries do the casting where necessary.

Bug: 400353541
Change-Id: I84db45bb35bbfb09f0d6bbde8f730d0f61f09c71
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6400633
Commit-Queue: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>