RPP: improve keyboard experience of annotations
This CL makes it so that when you tab into the annotation, you are
immediately placed into edit mode, and the cursor is put at the end of
the annotation.

This *sort of* worked for clicking with the mouse, but with the keyboard
you have to do a bit more work to get the cursor to display at the right
point. Because by default it will not appear.

Fixed: 406549702
Change-Id: I328d70763a09fc60a0c4258586fc97ba3d2f5658
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6401393
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>