Tooltip: ensure tooltip stays if focus moves from anchor into content
The tooltip has behaviour to ensure that when the focus
moves from the anchor to the tooltip that the tooltip does not hide,
however it only worked if the focus was the immediate child of the
tooltip. In our case (see attached bug) the focus was moving into a link
deeply nested within the contents of the tooltip.

This CL updates the code to make use of the blur target & relatedTarget
properties to detect this case and not hide the tooltip.

Fixed: 406543379
Change-Id: I7ad19fbde17e275663465e0bb86f8b2ea8ba768e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6402534
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>