RPP: make the perf panel UI inert when consent dialog shows
This CL makes the UI inert (it will not respond to user events) when the
dialog is showing, so you cannot tab around elements in the background.

We might want to expand this to a higher component so all of RPP is
covered, but this is enough for now.

Fixed: 406541486
Change-Id: Ie95ff549ba2392f6bdafeeff69c66e9ccd18ff7b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6403235
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>