RPP: improve visual look of annotation input with consent flow
When the consent flow popped up, the input lost focus and it renders as
a black rectangle (as it has no content). This CL updates it to maintain
the same visual state.

It also fixes a bug where focus was not properly restored; the early
exit condition in the setLabelEditabilityAndRemoveEmptyLabel function
meant that we wouldn't call focus, so I removed that. It's a pretty
cheap function that is not triggered that much so any theoretical boost
by early exiting is not worth it given it caused a bug.

Bug: 393063467
Change-Id: Id377a6ebee5b4ce0d4649e09f8fe7cbd5ed8d128
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6403241
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>