[RPP] Add LH Treemap button to Duplicated JavaScript insight
This lifts a bunch of code from Lighthouse[1] needed for opening
the Lighthouse Treemap webapp[2], a treemap visualization tool
for JS bundles. It shows all the scripts found on the page,
augmented with source map and any other data we send it (like
what large modules are duplicated on the page).

A couple refactors were necessary:

- The DuplicatedJavaScript model has a few new properties
  needed to assemble the data for the treemap
- computeGeneratedFileSizes moved from ScriptDuplication to
  ScriptsHandler, so it can be cached
- Add `inline` property to Script trace model

[1] https://github.com/GoogleChrome/lighthouse/blob/04f43865487b77384d5bfb02f774397ead5e3db2/core/audits/script-treemap-data.js
[1] https://github.com/GoogleChrome/lighthouse/blob/04f43865487b77384d5bfb02f774397ead5e3db2/report/renderer/open-tab.js#L102C10-L102C18
[2] https://googlechrome.github.io/lighthouse/treemap/

Bug: 394373632
Change-Id: I1c873e298d8ff0611f325a1752110d7682182f5e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6398204
Commit-Queue: Connor Clark <cjamcl@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>