[Workspace] Remove resetting working copy after text change
We started using `setWorkingCopy` to update UISourceCodes
programmatically from other parts of DevTools (e.g. Styles tab & PatchWidget).

Because of this, the logic for comparing the cleanliness of the
code needs fixing and crrev.com/c/6381859 attempted doing it.
However, somehow, with that change; updating styles from styles tab
and calling `setWorkingCopy` do not update the `UISourceCode` at
all and there are test failures attached to that CL.

Until we figure out what's going on, it's best to disable the logic
for resetting working copy when `isClean()` reports true on text change.

There is one more place where we have a similar logic:
activating the pretty toggle when the `isClean()` reports true on
text change but that's a smaller bug that we can afford right now.

Bug: 402699175
Change-Id: Id175cfaeb44ffa7728f30b28893654d73af6dd8e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6405494
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>