[Memory] Release all animations before memory operations
This is a workaround to remove user facing issue of having detached
elements in the profiles because of the animations panel.

What happens is, we keep the animations for replay in the animations panel.
However, when a DOM node is removed; we still keep the
animation related to the DOM node in the animations panel and
this causes the node to be retained and show up in the memory
profiles as detached elements.

Fixed: 400635410
Change-Id: I686799c1069978fbb78250d88ae6459a5ee26572
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6406316
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>