[RPP] Sort sample events after running SamplesIntegrator
With https://crrev.com/c/6394268 events created for samples with a trace
id were added after the ones created for all the obtained samples. As a
consequence, the containing array ended up unsorted. This can cause
an error  where an unordered event is unexpectedly found when
building the tree hierarchy in the trace engine.

To fix this, the containing array is ordered at the end of the
buildProfileCalls, which contains the main implementation of the
SamplesIntegrator.

Bug: none
Change-Id: I9e26e7890b91f5be1510e4fbc7c29e577aed4ed0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6408651
Commit-Queue: Paul Irish <paulirish@chromium.org>
Auto-Submit: Andres Olivares <andoli@chromium.org>
Reviewed-by: Paul Irish <paulirish@chromium.org>