RPP: simplify / reduce frontend validity checks in ignore list
The validation checks in this component were a bit eager and showed a
lot when the user interacted. Or it was easy to put some invalid input
in, but then click away and it was unclear if the message should stay or
go (it stayed).

In the attached bug Paul made this suggestion:

> get rid of the user-facing validity (and ensure we dont mess up ignorelist settings by committing an empty rule)

I like this, and therefore this CL does that. It still checks the regex
is valid before storing it, because else the ignore list errors, but now
we don't show anything to the user.

I think this is OK because it's fairly obvious an empty regex is not
valid. The only edge case is if you've typed a complex regex in and not
closed everything properly, so we might want to consider that, but I
don't think many users use it in this way so for now let's make our
validation less noisy.

Bug: 394325644
Change-Id: Ib16a62cbb17a4bb3852f6f7f2516ad7e85f7a1bc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6415795
Reviewed-by: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>