Use CPU for canvases that primarily read back
By enabling the willReadFrequently option when getting the canvas'
context, the CPU will be used for rendering. When there are only few
things drawn on the canvas before a readback happens, this has a
performance improvement. Otherwise the canvas contents would need to be
rendered in the GPU and then copied back over to the CPU.
Additionally, with the noise that we are adding to canvas readbacks
(https://github.com/explainers-by-googlers/canvas-noise-on-readbacks)
the results might get noised if they are drawn on the GPU.

Bug: 377325952
Change-Id: Ia09f34eba21817479971731e4fd7c554afdeedc6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6425762
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>
Auto-Submit: Tom Van Goethem <tov@chromium.org>