[Quick open] Suggestion text is only set from outside
This changes the functionality of the suggestion text. The suggestion input is no longer set to be exactly like the input text on input. Instead, the suggestion text is only set from outside. If the suggestion text is not different from the input, there is no gain in showing anything anyway.

This fixes a bug where we'd have overlapping suggestion & input texts if the input text was too long. In that case, the input text was scrolled to the end, while the suggestion text was not.

Fixed: 405350055
Change-Id: I4933cb8f3c70790dd6885de744cc64369904b7b7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6428565
Reviewed-by: Kateryna Prokopenko <kprokopenko@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>