[AiAgent] Fix not saving requests resulting in function call for eval
If `#aidaFetch` results in a functionCall, it means the fetch
loop is already finished (we directly break afterwards).

However, when it yields the function call,
the `run` method breaks from the async generator loop
causing rest of the `#aidaFetch` method not to run.

This CL removes the `break` from `run` method so that
the rest of the `#aidaFetch` will continue to run.

Bug: none
Change-Id: Ife7adf28c944d9dff5844d48a72e0cbac86739f7
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6432430
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Ergün Erdoğmuş <ergunsh@chromium.org>