AI: add an aria-description to the chat input
We had feedback from the a11y tester that jumping the user into the
textarea could be confusing rather than focusing on the first element in
the panel and letting the user tab through. To avoid that confusion, we
agreed that adding a description to the textarea to tell the user about
the suggested prompts.

Fixed: 408289708
Change-Id: Ie1bc5b22c4f694628ed5ee339b32ca6ab819efa1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6433968
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>