RPP: store pure function calls as facts
This CL implements some basic caching for the pure functions that we
expose to the AI. This avoids them continually recalling the same
functions which is inefficient and also a poor user experience to see
the AI continually repeat the exact same calls.

The data is cached based on the insight, ensuring that if the insight
changes, we invalidate the cache.

We could do this better by caching based on the bounds of the insight
(because if the bounds are the same, we don't need to recall the
function, regardless of which insight it is) but that will be harder for
the AI to understand, whereas this works reliably based on my testing.

Bug: 408172181
Change-Id: I243becee05140a2c4ca148cb3aed25936285f29d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6439198
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>