[eslint] Fix issue with JsLog rule and VSCode
Remove the update if the Set didn't change during the rule run.
This make the rule take so little time it's no longer in the top 10
for time consumed.

Also conditionally reload the rule while if needed.
This can happen if the user checkouts a separate branch or commit,
this meant that the Set was out of sync and value could be deleted
unintentionally.

Bug: none
Change-Id: I29ff66e4e9f979444e5a22f1f9c001af9769dfc5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6439513
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Auto-Submit: Nikolay Vitkov <nvitkov@chromium.org>
Commit-Queue: Nikolay Vitkov <nvitkov@chromium.org>