[sources] Show connectable folders inline in the Workspace tab.
Per UX deck[^1], the connectable Automatic Workspace folders must be
shown inline in the Workspace tab (of the Sources panel) with a button
next to it, that allows the user to connect the folder.

This is modeled by adding a transient `Workspace.Workspace.Project` as a
placeholder for the automatic file system, while it's not connected yet.
The placeholder project automatically disappears when the folder is
connected.

Tests will be added in a separate CL.

[^1]: http://go/chrome-devtools:automatic-workspace-folders-ux

Bug: 404170628
Change-Id: I2dd7795e5fd866c605474cd09cc41df3d3d3a2ae
Screenshot: http://screen/BpHS5AmayeQ6maG.png
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6440430
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>