[PP] Handle multiple async task runs caused by the same schedule event
Before this CL we assumed that every async task run event was connected
via Perfetto's flow API to its corresponding scheduling event. However,
because for async tasks the flow id is created based on its memory
address, if a single scheduling task causes multiple "run" events, each
of them  will appear as parent of the subsequent one (since they
will all have the same flow id).

To fix this, when we detect a schedule event to be the async parent of
a run event, we cache it. This way, when we detect the same run event to
be the async parent of another run, we can assign the original schedule
event as its effective async parent event

This fixes a bug in itself and will also be used to improve how we build
async stack traces for trace entries (that will happen in a follow up).

Bug: 389056697
Change-Id: I8234dce05709f6e6c355613065d4240676eb4496
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6448277
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>