[build] Add 'copy_to_gen_with_grd' GN template
This CL adds a new GN template that works similar to 'copy_to_gen'
but also marks the file as included in the release GRD bundle.

This makes it also similar to the 'devtools_pre_built' but not
restricted to JS files.

Drive-by: Split the CSS color tokens off the web test runner. We are
looking into excluding the integration_test_runner.html from the
release GRD and having the CSS color tokens woven together with the
web test runner in the BUILD.gn prevents that.

R=bmeurer@chromium.org

Bug: 382387670
Change-Id: Ie1399aa07e44b418fe295eb29b1d5f609934d48f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6450344
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>