[PP] Rename stack trace helper
getZeroIndexedStackTraceForEvent->getZeroIndexedStackTraceInEventPayload

To make it clear that it returns the stack trace contained in the
payload of a trace event. This way it's easier to distinguish from the
StackTraceForEvent.ts helper, which can build async stack traces using
the call hierarchy built by the RendererHandler.

Bug: 389056697
Change-Id: I171d2d2f955a6c78f87f8ce849af9ea4bf7258fc
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6449137
Reviewed-by: Nancy Li <nancyly@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>