[PP] Generalize StackTraceForEvent helper
Such that we look for the nearest profile call ancestor of any given
event and use its stack trace as the event's own stack. This is instead
of only supporting profile calls and an allow list of event types.

Drive-by: fix extension entry case to ensure console.timeStamp extension
entries are supported by the helper.

Bug: 389056697
Change-Id: I4f1bda365ee9b42ed8a4301dfb0bffd7b5bd0663
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6450704
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>