[PP] Support async stack traces for events triggered by async js calls
In an ideal world, for every single async JS call, we are able to connect the calling JS function (scheduler) to the function being called/scheduled. However, due to limitations with sampling, this is not always possible because we might not even receive any information from the JS calls in either ends of the async task  (scheduler - scheduled). In these cases, the call would be completely missing in the built call graph.

Nonetheless, the trace events tracking the browser scheduling of async tasks, are always present. These events provide data about the async task in question, like its name (e.g. setTimeout, setInterval, etc.). For this reason, this CL introduces a fallback behaviour for cases in which either JS call isn't present.

Two changes implement this behaviour:

1. In AsyncJSCallsHandler: We now have a separate map that's filled regardless of the presence of either JS call in the async task. We fall back to the corresponding tracking event (AsyncTaskScheduled or AsyncTaskRun) if the JS call at that end of the task isn't found.

2. In StackTraceForEvent: When going up in the trace event hierarchy to rebuild stack traces, if we detect that a trace event has an asynchronous parent (based on change #1) we move to that parent's stack and add the async task name to the async stack description. This differs to the previous behaviour, where we would only move across async parents from JS calls. This change is basically a generalization of the old behaviour.

Bug: 389056697
Change-Id: Ic9154ae8a2741afe243474f4184d2c51b5c6af09
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6449138
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>