RPP: improve Insights prompt for LCP
This CL updates the prompt with some specific LCP advice to improve
recommendations. It's slightly unfortunate that these have to go into
the preamble (because they take up space even for non-LCP insights) but
in my experimentation so far the information is much more likely to be
followed if we put them there. Given that right now most of the Insights
that support AI are LCP based, I'm not too concerned, and we can explore
this further as we add support for more insights.

R=alinavarkki@chromium.org

Bug: 409706097
Change-Id: I8c49cb7a48e447ba768af90d83e03115a4165991
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6455427
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Alina Varkki <alinavarkki@chromium.org>
Reviewed-by: Alina Varkki <alinavarkki@chromium.org>