[cleanup] Greatly simplify `ScheduledRender.scheduleRender`.
The previous implementation was unnecessary complex and added a lot of
overhead due to the heavy use of weak collections, and also required the
developer to explicitly create bound functions for the render methods.
In particular the weak collections were useless, since the closure in
the function was holding on to the objects anyways.

All of this complexity is unnecessary, and we can just use a very simple
implementation instead, without any weak collections, and without
forcing the user of the helper to create bound functions.

Bug: 40182218
Change-Id: I30a5ad5abe0e303f90919f83658d494e4a84a122
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6461926
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Danil Somsikov <dsv@chromium.org>