SpeculationRules: Avoid rendering non-existent ruleset
As the issue describes, when a developer selects one of speculation
rules ruleset on the dropdown in the preloading view and then reloads
the current page, the view becomes blank. This is because the dropdown
keeps referring to the ruleset that no longer exists and fails to render
the non-existent item. The same issue happens when a ruleset is removed
from the page, while the view is referring to the ruleset.

This CL fixes it by resetting the selected item if the corresponding
ruleset no longer exists on the model update event.

Bug: 41492925
Change-Id: Ib27a6394b07c7a7774e4b0f868952cb20aa42106
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6472075
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Ken Okada <kenoss@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>