[MenuButton] Add a menu button component
This introduces a menu button component that opens a menu. Native menus
that are open via click may cause issues with screen readers (JAWS),
where the focus of the screen reader is on the background instead of the
menu itself. The cause is a timing issue, in which two focus events
clash (one on the button, and one on the context menu). This menu button
delays the triggering of the menu, in order to avoid focus issues.

Fixed: 378461639
Change-Id: I01db4f80415cf8ca2017b5a3fdbc4dd4831480d4
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6467165
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>