Do not surface third-party cookie phaseout related issues.
Since there is no longer any notion of any ensuing third-party
cookie phaseout in chrome, we should not surface the issues and
console messages that hint different future behavior in chrome.

Instead of completely purging these cookie issues, we will prevent
these issues from being aggregated and then surfaced in the issues
pane. This is done because the "Third-party cookies" report in the
privacy and security panel relies on these issues for its table.

In the future, we ought to have a more specialized solution to
provide the report with its data. See this bug for progress:
http://crbug.com/413300867

This change also removes the third-party cookie console warning as
shown in the screenshot below. Note that a console message will
still be shown when third-party cookies are blocked (due to user
settings or incognito).

Before (left) and after (right):
https://screenshot.googleplex.com/8oYdJ2UzTvSxvwh.png

Bug: 413285149
Change-Id: I9202e091514cf0493aa84af108e3644281e39e98
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/6488309
Reviewed-by: Danil Somsikov <dsv@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Joshua Thomas <masnoble@chromium.org>